:recycle: Refactor for generate commands
